service/requirements.txt
albumentations==1.3.0
numpy==1.23.1
opencv_contrib_python==4.6.0.66
pandas==1.4.3
Pillow==9.3.0
streamlit==1.14.1
torch==1.12.0
